requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
astroid==2.4.0
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dateparser==0.7.4
discord.py==1.3.3
entrypoints==0.3
flake8==3.7.9
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pathspec==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.5.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
tzlocal==2.1b1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
